+++ packages/package/test/fixtures/javascript/expected/Test.svelte.d.ts
@@ -36,5 +36,7 @@ declare const __propDef: {
 			astring: string;
 		};
 	};
+	exports?: {};
+	bindings?: string;
 };
 export {};